Adding plasma tendrils to game scene
We're adding a PlasmaField object which contains a collection o
Tendril objects. Each Tendril is a Bezier curve and has a bunch of
tweens on it to make it wibble about nicely.

The PlasmaField manages the Tendrils and also draws them (this lets us
do drawing while knowing about all the tendrils at once).

We can click to 'fire' the beam, clicking again will reset the
tendrils to their previous position and behaviour. It's a lot of
tweening nonsense doing this so it'll be good to get straight exactly
what we want the behaviour to be so we can lock it down and then
ignore it forever.