Add tendril->enemy collision
The tendrils all point at the same angle while firing, and the enemy
ships get spawned at  random angle and move straight inwards, so we
really only need to check if the spawn angle is close to the current
firing angle of one of the tendrils. We add a buffer either side of 15
degrees so we end up affecting a cone of enemy ships.

Also turning off the shield while firing lets the ships get close to
the space station which is fun and looks super cool.